refactor eval util and testing
