+++ training/src/main/scala/com/airbnb/aerosolve/training/pipeline/EvalUtil.scala
@@ -15,26 +15,6 @@ import scala.collection.JavaConverters._
 object EvalUtil {
   val log: Logger = LoggerFactory.getLogger(this.getClass)
 
-  def scoreExamples(
-      sc: SparkContext,
-      transformer: Transformer,
-      modelOpt: AbstractModel,
-      examples: RDD[Example],
-      isTraining: Example => Boolean,
-      labelKey: String): RDD[(Float, String)] = {
-    val modelBC = sc.broadcast(modelOpt)
-    val transformerBC = sc.broadcast(transformer)
-    val scoreAndLabel = examples
-      .map(example => {
-        transformerBC.value.combineContextAndItems(example)
-        val score = modelBC.value.scoreItem(example.example.get(0))
-        val rank = example.example.get(0).floatFeatures.get(labelKey).get("")
-        val label = (if (isTraining(example)) "TRAIN_" else "HOLD_") + (if (rank > 0) "P" else "N")
-        (score, label)
-      })
-    scoreAndLabel
-  }
-
   def scoreExamplesForEvaluation(
       sc: SparkContext,
       transformer: Transformer,
@@ -111,64 +91,5 @@ object EvalUtil {
     result.setScore(prob)
     result.setLabel(rank)
     result
-  }
-
-  def getClassificationAUC(records : Seq[EvaluationRecord]) : Double = {
-    // Find minimal and maximal scores
-    var minScore = records.head.score
-    var maxScore = minScore
-
-    records.foreach(record => {
-      val score = record.score
-
-      minScore = Math.min(minScore, score)
-      maxScore = Math.max(maxScore, score)
-    })
-
-    if (minScore >= maxScore) {
-      log.warn("max score smaller than or equal to min score (%f, %f).".format(minScore, maxScore))
-      maxScore = minScore + 1.0
-    }
-
-    // For AUC evaluation
-    val buckets = records
-      .map(x => evaluateRecordForAUC(x, minScore, maxScore))
-      .groupBy(_._1)
-      .map(x => x._2.reduce((a, b) => (a._1, (a._2._1 + b._2._1, a._2._2 + b._2._2))))
-      .toArray
-      .sortBy(x => x._1)
-
-    getAUC(buckets.map(x => (x._2._1, x._2._2)))
-  }
-
-  private def evaluateRecordForAUC(
-      record : EvaluationRecord,
-      minScore : Double,
-      maxScore : Double) : (Long, (Long, Long)) = {
-    var offset = 0
-
-    if (record.label <= 0) {
-      offset += 1
-    }
-
-    val score : Long = ((record.score - minScore) / (maxScore - minScore) * 100).toLong
-
-    offset match {
-      case 0 => (score, (1, 0))
-      case 1 => (score, (0, 1))
-    }
-  }
-
-  private def getAUC(buckets: Array[(Long, Long)]): Double = {
-    val tot = buckets.reduce((x, y) => (x._1 + y._1, x._2 + y._2))
-    var auc = 0.0
-    var cs=(0L, 0L)
-    for (x <- buckets) {
-      // area for the current slice: (TP0+TP1)/2*(TN1-TN0)
-      auc += ((tot._1 - cs._1) + (tot._1 - cs._1 - x._1)) / 2.0 * x._2
-      // (TP0, TN0) -> (TP1, TN1)
-      cs = (cs._1 + x._1, cs._2 + x._2)
-    }
-    auc / tot._1 / tot._2
   }
 }
